refactor(audit): split the 2.3k-line CLI into focused modules
src/cli/audit.js is now a thin command shell. Internals live in
src/audit/{shared,args,config,scan,contract,baseline,report,
render-text,render-markdown,render-html,render-github,render-utils}.js.
No behaviour change; the public audit API and CLI exports are the
same objects. Module graph is acyclic.
